refactor(ai): drop motion + revert viz provider lift; fix artifact open/close
- Replace the framer/motion-based morph between AiArtifactButton and the
  floating panel with a vanilla CSS scale + fade-in on the panel. The
  morph was misbehaving when switching between artifacts (layoutId
  collision: the new panel and the appearing-button both claimed the
  same id in opposite directions) and re-animating on unrelated layout
  shifts like sidebar toggle.
- Drop the AiVisualizationProviders abstraction: providers move back
  into AiVisualizationRenderer. The floating panel's head (which uses
  useVisualizationContext via AiChartQuickOptions) now rides in the
  renderer's existing headerContent slot, so the head sits inside the
  provider tree naturally. Pill stays as a positioned overlay sibling
  since the switcher itself doesn't read viz context.
- Remove LayoutGroup from AiAgentPageLayout and the motion dep from
  package.json + lockfile.
- Fix artifact button click behavior: clicking the open artifact's
  button now closes it (clearArtifact), enabling toggle-to-close.
  Switching artifacts works for free now that there's no shared-id
  morph stalling the panel re-render.
