+++ packages/frontend/package.json
@@ -103,7 +103,6 @@
         "monaco-editor": "0.44.0",
         "monaco-sql-languages": "0.12.2",
         "monaco-yaml": "5.4.0",
-        "motion": "12.38.0",
         "polished": "4.2.2",
         "postcss-preset-mantine": "1.18.0",
         "postcss-simple-vars": "7.0.1",
