+++ packages/frontend/src/ee/features/aiCopilot/components/ChatElements/AiVisualizationRenderer.tsx
@@ -1,49 +1,106 @@
 import {
     AiResultType,
+    ChartType,
+    ECHARTS_DEFAULT_COLORS,
     getGroupByDimensions,
     getWebAiChartConfig,
     type AiAgentChartTypeOption,
     type ApiAiAgentThreadMessageVizQuery,
+    type ChartConfig,
+    type EChartsSeries,
     type ToolRunQueryArgs,
     type ToolTableVizArgs,
     type ToolTimeSeriesArgs,
     type ToolVerticalBarArgs,
 } from '@lightdash/common';
-import { Box, Center, Group, Stack, Text } from '@mantine-8/core';
+import {
+    Box,
+    Center,
+    Group,
+    Stack,
+    Text,
+    useMantineColorScheme,
+} from '@mantine-8/core';
 import { IconExclamationCircle } from '@tabler/icons-react';
-import { useMemo, type FC, type ReactNode } from 'react';
+import { useCallback, useMemo, useState, type FC, type ReactNode } from 'react';
+import { useParams } from 'react-router';
 import MantineIcon from '../../../../../components/common/MantineIcon';
+import { SeriesContextMenu } from '../../../../../components/Explorer/VisualizationCard/SeriesContextMenu';
 import LightdashVisualization from '../../../../../components/LightdashVisualization';
+import VisualizationProvider from '../../../../../components/LightdashVisualization/VisualizationProvider';
+import { DrillDownModal } from '../../../../../components/MetricQueryData/DrillDownModal';
+import MetricQueryDataProvider from '../../../../../components/MetricQueryData/MetricQueryDataProvider';
+import UnderlyingDataModal from '../../../../../components/MetricQueryData/UnderlyingDataModal';
+import { type EchartsSeriesClickEvent } from '../../../../../components/SimpleChart';
 import ErrorBoundary from '../../../../../features/errorBoundary/ErrorBoundary';
+import { useProjectColorPalette } from '../../../../../hooks/appearance/useProjectColorPalette';
 import useHealth from '../../../../../hooks/health/useHealth';
+import { useExplore } from '../../../../../hooks/useExplore';
+import { type InfiniteQueryResults } from '../../../../../hooks/useQueryResults';
 import { AgentVisualizationChartTypeSwitcher } from './AgentVisualizationChartTypeSwitcher';
 import AgentVisualizationFilters from './AgentVisualizationFilters';
 import AgentVisualizationMetricsAndDimensions from './AgentVisualizationMetricsAndDimensions';
 
 type Props = {
     vizQueryData: ApiAiAgentThreadMessageVizQuery;
+    results: InfiniteQueryResults;
     chartConfig:
         | ToolTableVizArgs
         | ToolTimeSeriesArgs
         | ToolVerticalBarArgs
         | ToolRunQueryArgs;
     selectedChartType: AiAgentChartTypeOption | null;
-    // When provided, an inline switcher is shown above the chart. Omit it
-    // (e.g. on the floating panel) when a parent renders its own switcher.
+    // When provided, an inline switcher is rendered above the chart. Omit
+    // it (e.g. on the floating panel) when a parent renders its own.
     onChartTypeChange?: (type: AiAgentChartTypeOption) => void;
+    // Forwarded when the underlying viz expands its config — used by the
+    // dashboard path to sync the change back into the query cache.
+    onExpandedChartConfigChange?: (config: ChartConfig) => void;
     headerContent?: ReactNode;
 };
 
 export const AiVisualizationRenderer: FC<Props> = ({
     vizQueryData,
+    results,
     chartConfig,
     selectedChartType,
     onChartTypeChange,
+    onExpandedChartConfigChange,
     headerContent,
 }) => {
     const { data: health } = useHealth();
+    const { projectUuid } = useParams<{ projectUuid: string }>();
+    const { data: resolvedPalette } = useProjectColorPalette(projectUuid);
+    const { colorScheme } = useMantineColorScheme();
 
-    const { metricQuery, fields } = vizQueryData.query;
+    const colorPalette = useMemo(() => {
+        if (colorScheme === 'dark' && resolvedPalette?.darkColors) {
+            return resolvedPalette.darkColors;
+        }
+        return resolvedPalette?.colors ?? ECHARTS_DEFAULT_COLORS;
+    }, [colorScheme, resolvedPalette]);
+
+    const { metricQuery, fields, resolvedTimezone } = vizQueryData.query;
+    const tableName = metricQuery?.exploreName;
+    const { data: explore } = useExplore(tableName);
+
+    const [echartsClickEvent, setEchartsClickEvent] =
+        useState<EchartsSeriesClickEvent | null>(null);
+    const [echartsSeries, setEchartsSeries] = useState<EChartsSeries[]>([]);
+
+    const [expandedChartConfig, setExpandedChartConfig] = useState<
+        ChartConfig | undefined
+    >(undefined);
+
+    const resultsData = useMemo(
+        () => ({
+            ...results,
+            metricQuery,
+            fields,
+            resolvedTimezone: resolvedTimezone ?? undefined,
+        }),
+        [results, metricQuery, fields, resolvedTimezone],
+    );
 
     const webAiChartConfig = useMemo(
         () =>
@@ -77,6 +134,14 @@ export const AiVisualizationRenderer: FC<Props> = ({
             ? (webAiChartConfig.vizTool.chartConfig?.defaultVizType ?? 'table')
             : 'table';
 
+    const handleChartConfigChange = useCallback(
+        (newConfig: ChartConfig) => {
+            setExpandedChartConfig(newConfig);
+            onExpandedChartConfigChange?.(newConfig);
+        },
+        [onExpandedChartConfigChange],
+    );
+
     if (!webAiChartConfig.echartsConfig) {
         return (
             <Center h={300}>
@@ -91,53 +156,104 @@ export const AiVisualizationRenderer: FC<Props> = ({
     }
 
     return (
-        <Stack gap="md" h="100%">
-            {headerContent}
-            {webAiChartConfig.type === AiResultType.QUERY_RESULT &&
-                onChartTypeChange && (
-                    <Group justify="flex-end">
-                        <AgentVisualizationChartTypeSwitcher
-                            metricQuery={metricQuery}
-                            selectedChartType={
-                                selectedChartType ?? defaultChartType
-                            }
-                            hasGroupByDimensions={
-                                (groupByDimensions?.length ?? 0) > 0
-                            }
-                            onChartTypeChange={onChartTypeChange}
-                        />
-                    </Group>
-                )}
-            <Box
-                flex="1"
-                style={{
-                    // Scrolling for tables
-                    overflow: 'auto',
+        <MetricQueryDataProvider
+            metricQuery={metricQuery}
+            tableName={tableName}
+            explore={explore}
+            queryUuid={vizQueryData.query.queryUuid}
+            resolvedTimezone={resolvedTimezone}
+        >
+            <VisualizationProvider
+                key={selectedChartType ?? 'default'}
+                resultsData={resultsData}
+                chartConfig={
+                    expandedChartConfig ?? webAiChartConfig.echartsConfig
+                }
+                parameters={vizQueryData.query.usedParametersValues}
+                columnOrder={[
+                    ...metricQuery.dimensions,
+                    ...metricQuery.metrics,
+                    ...metricQuery.tableCalculations.map((tc) => tc.name),
+                ]}
+                pivotTableMaxColumnLimit={
+                    health?.pivotTable.maxColumnLimit ?? 60
+                }
+                initialPivotDimensions={groupByDimensions}
+                colorPalette={colorPalette}
+                isLoading={resultsData.isFetchingRows}
+                onSeriesContextMenu={(
+                    e: EchartsSeriesClickEvent,
+                    series: EChartsSeries[],
+                ) => {
+                    setEchartsClickEvent(e);
+                    setEchartsSeries(series);
                 }}
+                onChartConfigChange={handleChartConfigChange}
+                unsavedMetricQuery={metricQuery}
             >
-                <LightdashVisualization
-                    className="sentry-block ph-no-capture"
-                    data-testid="ai-visualization"
-                />
-            </Box>
-
-            <Stack gap="xs">
-                <ErrorBoundary>
-                    {displayMetricsAndDimensions && (
-                        <AgentVisualizationMetricsAndDimensions
-                            metricQuery={metricQuery}
-                            fieldsMap={fields}
+                <Stack gap="md" h="100%">
+                    {headerContent}
+                    {webAiChartConfig.type === AiResultType.QUERY_RESULT &&
+                        onChartTypeChange && (
+                            <Group justify="flex-end">
+                                <AgentVisualizationChartTypeSwitcher
+                                    metricQuery={metricQuery}
+                                    selectedChartType={
+                                        selectedChartType ?? defaultChartType
+                                    }
+                                    hasGroupByDimensions={
+                                        (groupByDimensions?.length ?? 0) > 0
+                                    }
+                                    onChartTypeChange={onChartTypeChange}
+                                />
+                            </Group>
+                        )}
+                    <Box
+                        flex="1"
+                        style={{
+                            // Scrolling for tables
+                            overflow: 'auto',
+                        }}
+                    >
+                        <LightdashVisualization
+                            className="sentry-block ph-no-capture"
+                            data-testid="ai-visualization"
                         />
-                    )}
 
-                    {chartConfig.filters ? (
-                        <AgentVisualizationFilters
-                            filters={metricQuery.filters}
-                            fieldsMap={fields}
-                        />
-                    ) : null}
-                </ErrorBoundary>
-            </Stack>
-        </Stack>
+                        {webAiChartConfig.echartsConfig.type ===
+                            ChartType.CARTESIAN && (
+                            <SeriesContextMenu
+                                echartsSeriesClickEvent={
+                                    echartsClickEvent ?? undefined
+                                }
+                                dimensions={metricQuery.dimensions}
+                                series={echartsSeries}
+                                explore={explore}
+                            />
+                        )}
+                        <UnderlyingDataModal />
+                        <DrillDownModal />
+                    </Box>
+
+                    <Stack gap="xs">
+                        <ErrorBoundary>
+                            {displayMetricsAndDimensions && (
+                                <AgentVisualizationMetricsAndDimensions
+                                    metricQuery={metricQuery}
+                                    fieldsMap={fields}
+                                />
+                            )}
+
+                            {chartConfig.filters ? (
+                                <AgentVisualizationFilters
+                                    filters={metricQuery.filters}
+                                    fieldsMap={fields}
+                                />
+                            ) : null}
+                        </ErrorBoundary>
+                    </Stack>
+                </Stack>
+            </VisualizationProvider>
+        </MetricQueryDataProvider>
     );
 };
